version consistency: generalize MultiMzSqlExecutors
